requirements.txt
gradio
pandas
biopython
torch